+++ packages/util-crypto/src/address/keyMulti.spec.ts
@@ -2,21 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import createKeyMulti, { sortU8a } from './keyMulti';
-
-describe('sortU8a', (): void => {
-  it('sorts a simple set of u8a', (): void => {
-    expect(
-      sortU8a([new Uint8Array([1, 2, 3]), new Uint8Array([4, 5, 6]), new Uint8Array([2, 3, 4])])
-    ).toEqual([new Uint8Array([1, 2, 3]), new Uint8Array([2, 3, 4]), new Uint8Array([4, 5, 6])]);
-  });
-
-  it('sorts a simple set of u8a (not the same lengths)', (): void => {
-    expect(
-      sortU8a([new Uint8Array([1, 2, 3, 4]), new Uint8Array([4, 5, 6]), new Uint8Array([1, 2, 3, 5])])
-    ).toEqual([new Uint8Array([1, 2, 3, 4]), new Uint8Array([1, 2, 3, 5]), new Uint8Array([4, 5, 6])]);
-  });
-});
+import { createKeyMulti } from '.';
 
 describe('createKeyMulti', (): void => {
   it('creates a valid multikey (aligning with Rust, needs sorting)', (): void => {