+++ packages/util/src/u8a/index.ts
@@ -8,6 +8,7 @@
 
 export { default as u8aConcat } from './concat';
 export { default as u8aFixLength } from './fixLength';
+export { default as u8aSorted } from './sorted';
 export { default as u8aToBn } from './toBn';
 export { default as u8aToBuffer } from './toBuffer';
 export { default as u8aToHex } from './toHex';